requirements.txt
torch>=1.12.0
tensorflow>=2.9.0
opencv-python
numpy
pandas
flask
pyserial
pymavlink